Factor deserialization out of Field
